Moved AFont, AColor to AdmiralTheme + ColorPalette, FontPalette
