+++ Source/AdmiralTheme/Classes/CoreFoundation/AppTheme/Colors/ColorPalette.swift
@@ -4,289 +4,299 @@
 //
 //  Generated by Fastlane Figma Plugin.
 //
-
 import UIKit
 import SwiftUI
-
 // swiftlint:disable all
-
+/// The value that sets style for dark or light mode
 public enum ColorPaletteStyle: String, Codable {
+    /// The light mode
     case light
+    /// The dark mode
     case dark
 }
+/**
+ ColorPalette - The container that include a list of colors.
+ You can create your own ColorPalette or  expand the current color palette by using the mechanism of inheritance.
+ ## Example to create ColorPalette:
+ # Code
+ let colorPalette = ColorPalette(...)
+ class CustomColorPalatte: ColorPalette {}
+ ```
+ */
+/// The container that include a list of colors.
+open class ColorPalette: Codable {
 
-// MARK: - ColorPalette
+    // MARK: - Public Properties
 
-public class ColorPalette: Codable {
-        
     public var style: ColorPaletteStyle
     public var disabledAlpha: CGFloat
-     
-    public var backgroundAccent: AColor 
-    public var backgroundAccentDark: AColor 
-    public var backgroundAccentPressed: AColor 
-    public var backgroundAdditionalOne: AColor 
-    public var backgroundAdditionalOnePressed: AColor 
-    public var backgroundAdditionalTwo: AColor 
-    public var backgroundAdditionalTwoPressed: AColor 
-    public var backgroundAttention: AColor 
-    public var backgroundAttentionPressed: AColor 
-    public var backgroundBasic: AColor 
-    public var backgroundError: AColor 
-    public var backgroundErrorPressed: AColor 
-    public var backgroundExtraSurface: AColor 
-    public var backgroundModalControl: AColor 
-    public var backgroundModalView: AColor 
-    public var backgroundSecondary: AColor 
-    public var backgroundSelected: AColor 
-    public var backgroundSelectedPressed: AColor 
-    public var backgroundShadow: AColor 
-    public var backgroundSuccess: AColor 
-    public var backgroundSuccessPressed: AColor 
-    public var elementAccent: AColor 
-    public var elementAccentAdditional: AColor 
-    public var elementAccentPressed: AColor 
-    public var elementAdditional: AColor 
-    public var elementAttention: AColor 
-    public var elementAttentionDefault: AColor 
-    public var elementAttentionPressed: AColor 
-    public var elementBarDivider: AColor 
-    public var elementContrast: AColor 
-    public var elementError: AColor 
-    public var elementErrorDefault: AColor 
-    public var elementErrorPressed: AColor 
-    public var elementExtra: AColor 
+    public var backgroundAccent: AColor
+    public var backgroundAccentDark: AColor
+    public var backgroundAccentPressed: AColor
+    public var backgroundAdditionalOne: AColor
+    public var backgroundAdditionalOnePressed: AColor
+    public var backgroundAdditionalTwo: AColor
+    public var backgroundAdditionalTwoPressed: AColor
+    public var backgroundAttention: AColor
+    public var backgroundAttentionPressed: AColor
+    public var backgroundBasic: AColor
+    public var backgroundError: AColor
+    public var backgroundErrorPressed: AColor
+    public var backgroundExtraSurface: AColor
+    public var backgroundModalControl: AColor
+    public var backgroundModalView: AColor
+    public var backgroundSecondary: AColor
+    public var backgroundSelected: AColor
+    public var backgroundSelectedPressed: AColor
+    public var backgroundShadow: AColor
+    public var backgroundSuccess: AColor
+    public var backgroundSuccessPressed: AColor
+    public var elementAccent: AColor
+    public var elementAccentAdditional: AColor
+    public var elementAccentPressed: AColor
+    public var elementAdditional: AColor
+    public var elementAttention: AColor
+    public var elementAttentionDefault: AColor
+    public var elementAttentionPressed: AColor
+    public var elementBarDivider: AColor
+    public var elementContrast: AColor
+    public var elementError: AColor
+    public var elementErrorDefault: AColor
+    public var elementErrorPressed: AColor
+    public var elementExtra: AColor
     public var elementPrimary: AColor
     public var elementPrimaryPressed: AColor
-    public var elementSecondary: AColor 
-    public var elementStaticWhite: AColor 
-    public var elementSuccess: AColor 
-    public var elementSuccessDefault: AColor 
-    public var elementSuccessPressed: AColor 
-    public var specialExtra01: AColor 
-    public var specialExtra02: AColor 
-    public var specialExtra03: AColor 
-    public var specialExtra04: AColor 
-    public var specialExtra05: AColor 
-    public var specialExtra06: AColor 
-    public var specialExtra07: AColor 
-    public var specialExtra08: AColor 
-    public var specialExtra09: AColor 
-    public var specialExtra10: AColor 
-    public var specialExtra11: AColor 
-    public var specialExtra12: AColor 
-    public var specialExtra13: AColor 
-    public var specialExtra14: AColor 
-    public var specialExtra15: AColor 
-    public var specialExtra16: AColor 
-    public var specialExtra17: AColor 
-    public var specialExtra18: AColor 
-    public var specialExtra19: AColor 
-    public var specialExtra20: AColor 
-    public var specialExtra21: AColor 
-    public var specialExtra22: AColor 
-    public var specialExtra23: AColor 
-    public var specialExtra24: AColor 
-    public var specialExtra25: AColor 
-    public var specialExtra26: AColor 
-    public var specialExtra27: AColor 
-    public var textAccent: AColor 
-    public var textAccentAdditional: AColor 
-    public var textAccentPressed: AColor 
-    public var textAdditional: AColor 
-    public var textAttention: AColor 
-    public var textAttentionDefault: AColor 
-    public var textAttentionPressed: AColor 
-    public var textContrast: AColor 
-    public var textError: AColor 
-    public var textErrorDefault: AColor 
-    public var textErrorPressed: AColor 
-    public var textMask: AColor 
-    public var textPrimary: AColor 
-    public var textSecondary: AColor 
-    public var textStaticWhite: AColor 
-    public var textSuccess: AColor 
-    public var textSuccessDefault: AColor 
+    public var elementSecondary: AColor
+    public var elementStaticWhite: AColor
+    public var elementSuccess: AColor
+    public var elementSuccessDefault: AColor
+    public var elementSuccessPressed: AColor
+    public var specialExtra01: AColor
+    public var specialExtra02: AColor
+    public var specialExtra03: AColor
+    public var specialExtra04: AColor
+    public var specialExtra05: AColor
+    public var specialExtra06: AColor
+    public var specialExtra07: AColor
+    public var specialExtra08: AColor
+    public var specialExtra09: AColor
+    public var specialExtra10: AColor
+    public var specialExtra11: AColor
+    public var specialExtra12: AColor
+    public var specialExtra13: AColor
+    public var specialExtra14: AColor
+    public var specialExtra15: AColor
+    public var specialExtra16: AColor
+    public var specialExtra17: AColor
+    public var specialExtra18: AColor
+    public var specialExtra19: AColor
+    public var specialExtra20: AColor
+    public var specialExtra21: AColor
+    public var specialExtra22: AColor
+    public var specialExtra23: AColor
+    public var specialExtra24: AColor
+    public var specialExtra25: AColor
+    public var specialExtra26: AColor
+    public var specialExtra27: AColor
+    public var textAccent: AColor
+    public var textAccentAdditional: AColor
+    public var textAccentPressed: AColor
+    public var textAdditional: AColor
+    public var textAttention: AColor
+    public var textAttentionDefault: AColor
+    public var textAttentionPressed: AColor
+    public var textContrast: AColor
+    public var textError: AColor
+    public var textErrorDefault: AColor
+    public var textErrorPressed: AColor
+    public var textMask: AColor
+    public var textPrimary: AColor
+    public var textSecondary: AColor
+    public var textStaticWhite: AColor
+    public var textSuccess: AColor
+    public var textSuccessDefault: AColor
     public var textSuccessPressed: AColor
+
+    // MARK: - Initializer
 
     public init(
         style: ColorPaletteStyle,
-        disabledAlpha: CGFloat, 
-        backgroundAccent: AColor, 
-        backgroundAccentDark: AColor, 
-        backgroundAccentPressed: AColor, 
-        backgroundAdditionalOne: AColor, 
-        backgroundAdditionalOnePressed: AColor, 
-        backgroundAdditionalTwo: AColor, 
-        backgroundAdditionalTwoPressed: AColor, 
-        backgroundAttention: AColor, 
-        backgroundAttentionPressed: AColor, 
-        backgroundBasic: AColor, 
-        backgroundError: AColor, 
-        backgroundErrorPressed: AColor, 
-        backgroundExtraSurface: AColor, 
-        backgroundModalControl: AColor, 
-        backgroundModalView: AColor, 
-        backgroundSecondary: AColor, 
-        backgroundSelected: AColor, 
-        backgroundSelectedPressed: AColor, 
-        backgroundShadow: AColor, 
-        backgroundSuccess: AColor, 
-        backgroundSuccessPressed: AColor, 
-        elementAccent: AColor, 
-        elementAccentAdditional: AColor, 
-        elementAccentPressed: AColor, 
-        elementAdditional: AColor, 
-        elementAttention: AColor, 
-        elementAttentionDefault: AColor, 
-        elementAttentionPressed: AColor, 
-        elementBarDivider: AColor, 
-        elementContrast: AColor, 
-        elementError: AColor, 
-        elementErrorDefault: AColor, 
-        elementErrorPressed: AColor, 
-        elementExtra: AColor, 
+        disabledAlpha: CGFloat,
+        backgroundAccent: AColor,
+        backgroundAccentDark: AColor,
+        backgroundAccentPressed: AColor,
+        backgroundAdditionalOne: AColor,
+        backgroundAdditionalOnePressed: AColor,
+        backgroundAdditionalTwo: AColor,
+        backgroundAdditionalTwoPressed: AColor,
+        backgroundAttention: AColor,
+        backgroundAttentionPressed: AColor,
+        backgroundBasic: AColor,
+        backgroundError: AColor,
+        backgroundErrorPressed: AColor,
+        backgroundExtraSurface: AColor,
+        backgroundModalControl: AColor,
+        backgroundModalView: AColor,
+        backgroundSecondary: AColor,
+        backgroundSelected: AColor,
+        backgroundSelectedPressed: AColor,
+        backgroundShadow: AColor,
+        backgroundSuccess: AColor,
+        backgroundSuccessPressed: AColor,
+        elementAccent: AColor,
+        elementAccentAdditional: AColor,
+        elementAccentPressed: AColor,
+        elementAdditional: AColor,
+        elementAttention: AColor,
+        elementAttentionDefault: AColor,
+        elementAttentionPressed: AColor,
+        elementBarDivider: AColor,
+        elementContrast: AColor,
+        elementError: AColor,
+        elementErrorDefault: AColor,
+        elementErrorPressed: AColor,
+        elementExtra: AColor,
         elementPrimary: AColor,
         elementPrimaryPressed: AColor,
-        elementSecondary: AColor, 
-        elementStaticWhite: AColor, 
-        elementSuccess: AColor, 
-        elementSuccessDefault: AColor, 
-        elementSuccessPressed: AColor, 
-        specialExtra01: AColor, 
-        specialExtra02: AColor, 
-        specialExtra03: AColor, 
-        specialExtra04: AColor, 
-        specialExtra05: AColor, 
-        specialExtra06: AColor, 
-        specialExtra07: AColor, 
-        specialExtra08: AColor, 
-        specialExtra09: AColor, 
-        specialExtra10: AColor, 
-        specialExtra11: AColor, 
-        specialExtra12: AColor, 
-        specialExtra13: AColor, 
-        specialExtra14: AColor, 
-        specialExtra15: AColor, 
-        specialExtra16: AColor, 
-        specialExtra17: AColor, 
-        specialExtra18: AColor, 
-        specialExtra19: AColor, 
-        specialExtra20: AColor, 
-        specialExtra21: AColor, 
-        specialExtra22: AColor, 
-        specialExtra23: AColor, 
-        specialExtra24: AColor, 
-        specialExtra25: AColor, 
-        specialExtra26: AColor, 
-        specialExtra27: AColor, 
-        textAccent: AColor, 
-        textAccentAdditional: AColor, 
-        textAccentPressed: AColor, 
-        textAdditional: AColor, 
-        textAttention: AColor, 
-        textAttentionDefault: AColor, 
-        textAttentionPressed: AColor, 
-        textContrast: AColor, 
-        textError: AColor, 
-        textErrorDefault: AColor, 
-        textErrorPressed: AColor, 
-        textMask: AColor, 
-        textPrimary: AColor, 
-        textSecondary: AColor, 
-        textStaticWhite: AColor, 
-        textSuccess: AColor, 
-        textSuccessDefault: AColor, 
+        elementSecondary: AColor,
+        elementStaticWhite: AColor,
+        elementSuccess: AColor,
+        elementSuccessDefault: AColor,
+        elementSuccessPressed: AColor,
+        specialExtra01: AColor,
+        specialExtra02: AColor,
+        specialExtra03: AColor,
+        specialExtra04: AColor,
+        specialExtra05: AColor,
+        specialExtra06: AColor,
+        specialExtra07: AColor,
+        specialExtra08: AColor,
+        specialExtra09: AColor,
+        specialExtra10: AColor,
+        specialExtra11: AColor,
+        specialExtra12: AColor,
+        specialExtra13: AColor,
+        specialExtra14: AColor,
+        specialExtra15: AColor,
+        specialExtra16: AColor,
+        specialExtra17: AColor,
+        specialExtra18: AColor,
+        specialExtra19: AColor,
+        specialExtra20: AColor,
+        specialExtra21: AColor,
+        specialExtra22: AColor,
+        specialExtra23: AColor,
+        specialExtra24: AColor,
+        specialExtra25: AColor,
+        specialExtra26: AColor,
+        specialExtra27: AColor,
+        textAccent: AColor,
+        textAccentAdditional: AColor,
+        textAccentPressed: AColor,
+        textAdditional: AColor,
+        textAttention: AColor,
+        textAttentionDefault: AColor,
+        textAttentionPressed: AColor,
+        textContrast: AColor,
+        textError: AColor,
+        textErrorDefault: AColor,
+        textErrorPressed: AColor,
+        textMask: AColor,
+        textPrimary: AColor,
+        textSecondary: AColor,
+        textStaticWhite: AColor,
+        textSuccess: AColor,
+        textSuccessDefault: AColor,
         textSuccessPressed: AColor
     ) {
         self.style = style
         self.disabledAlpha = disabledAlpha
-         
-        self.backgroundAccent = backgroundAccent 
-        self.backgroundAccentDark = backgroundAccentDark 
-        self.backgroundAccentPressed = backgroundAccentPressed 
-        self.backgroundAdditionalOne = backgroundAdditionalOne 
-        self.backgroundAdditionalOnePressed = backgroundAdditionalOnePressed 
-        self.backgroundAdditionalTwo = backgroundAdditionalTwo 
-        self.backgroundAdditionalTwoPressed = backgroundAdditionalTwoPressed 
-        self.backgroundAttention = backgroundAttention 
-        self.backgroundAttentionPressed = backgroundAttentionPressed 
-        self.backgroundBasic = backgroundBasic 
-        self.backgroundError = backgroundError 
-        self.backgroundErrorPressed = backgroundErrorPressed 
-        self.backgroundExtraSurface = backgroundExtraSurface 
-        self.backgroundModalControl = backgroundModalControl 
-        self.backgroundModalView = backgroundModalView 
-        self.backgroundSecondary = backgroundSecondary 
-        self.backgroundSelected = backgroundSelected 
-        self.backgroundSelectedPressed = backgroundSelectedPressed 
-        self.backgroundShadow = backgroundShadow 
-        self.backgroundSuccess = backgroundSuccess 
-        self.backgroundSuccessPressed = backgroundSuccessPressed 
-        self.elementAccent = elementAccent 
-        self.elementAccentAdditional = elementAccentAdditional 
-        self.elementAccentPressed = elementAccentPressed 
-        self.elementAdditional = elementAdditional 
-        self.elementAttention = elementAttention 
-        self.elementAttentionDefault = elementAttentionDefault 
-        self.elementAttentionPressed = elementAttentionPressed 
-        self.elementBarDivider = elementBarDivider 
-        self.elementContrast = elementContrast 
-        self.elementError = elementError 
-        self.elementErrorDefault = elementErrorDefault 
-        self.elementErrorPressed = elementErrorPressed 
-        self.elementExtra = elementExtra 
+
+        self.backgroundAccent = backgroundAccent
+        self.backgroundAccentDark = backgroundAccentDark
+        self.backgroundAccentPressed = backgroundAccentPressed
+        self.backgroundAdditionalOne = backgroundAdditionalOne
+        self.backgroundAdditionalOnePressed = backgroundAdditionalOnePressed
+        self.backgroundAdditionalTwo = backgroundAdditionalTwo
+        self.backgroundAdditionalTwoPressed = backgroundAdditionalTwoPressed
+        self.backgroundAttention = backgroundAttention
+        self.backgroundAttentionPressed = backgroundAttentionPressed
+        self.backgroundBasic = backgroundBasic
+        self.backgroundError = backgroundError
+        self.backgroundErrorPressed = backgroundErrorPressed
+        self.backgroundExtraSurface = backgroundExtraSurface
+        self.backgroundModalControl = backgroundModalControl
+        self.backgroundModalView = backgroundModalView
+        self.backgroundSecondary = backgroundSecondary
+        self.backgroundSelected = backgroundSelected
+        self.backgroundSelectedPressed = backgroundSelectedPressed
+        self.backgroundShadow = backgroundShadow
+        self.backgroundSuccess = backgroundSuccess
+        self.backgroundSuccessPressed = backgroundSuccessPressed
+        self.elementAccent = elementAccent
+        self.elementAccentAdditional = elementAccentAdditional
+        self.elementAccentPressed = elementAccentPressed
+        self.elementAdditional = elementAdditional
+        self.elementAttention = elementAttention
+        self.elementAttentionDefault = elementAttentionDefault
+        self.elementAttentionPressed = elementAttentionPressed
+        self.elementBarDivider = elementBarDivider
+        self.elementContrast = elementContrast
+        self.elementError = elementError
+        self.elementErrorDefault = elementErrorDefault
+        self.elementErrorPressed = elementErrorPressed
+        self.elementExtra = elementExtra
         self.elementPrimary = elementPrimary
         self.elementPrimaryPressed = elementPrimaryPressed
-        self.elementSecondary = elementSecondary 
-        self.elementStaticWhite = elementStaticWhite 
-        self.elementSuccess = elementSuccess 
-        self.elementSuccessDefault = elementSuccessDefault 
-        self.elementSuccessPressed = elementSuccessPressed 
-        self.specialExtra01 = specialExtra01 
-        self.specialExtra02 = specialExtra02 
-        self.specialExtra03 = specialExtra03 
-        self.specialExtra04 = specialExtra04 
-        self.specialExtra05 = specialExtra05 
-        self.specialExtra06 = specialExtra06 
-        self.specialExtra07 = specialExtra07 
-        self.specialExtra08 = specialExtra08 
-        self.specialExtra09 = specialExtra09 
-        self.specialExtra10 = specialExtra10 
-        self.specialExtra11 = specialExtra11 
-        self.specialExtra12 = specialExtra12 
-        self.specialExtra13 = specialExtra13 
-        self.specialExtra14 = specialExtra14 
-        self.specialExtra15 = specialExtra15 
-        self.specialExtra16 = specialExtra16 
-        self.specialExtra17 = specialExtra17 
-        self.specialExtra18 = specialExtra18 
-        self.specialExtra19 = specialExtra19 
-        self.specialExtra20 = specialExtra20 
-        self.specialExtra21 = specialExtra21 
-        self.specialExtra22 = specialExtra22 
-        self.specialExtra23 = specialExtra23 
-        self.specialExtra24 = specialExtra24 
-        self.specialExtra25 = specialExtra25 
-        self.specialExtra26 = specialExtra26 
-        self.specialExtra27 = specialExtra27 
-        self.textAccent = textAccent 
-        self.textAccentAdditional = textAccentAdditional 
-        self.textAccentPressed = textAccentPressed 
-        self.textAdditional = textAdditional 
-        self.textAttention = textAttention 
-        self.textAttentionDefault = textAttentionDefault 
-        self.textAttentionPressed = textAttentionPressed 
-        self.textContrast = textContrast 
-        self.textError = textError 
-        self.textErrorDefault = textErrorDefault 
-        self.textErrorPressed = textErrorPressed 
-        self.textMask = textMask 
-        self.textPrimary = textPrimary 
-        self.textSecondary = textSecondary 
-        self.textStaticWhite = textStaticWhite 
-        self.textSuccess = textSuccess 
-        self.textSuccessDefault = textSuccessDefault 
+        self.elementSecondary = elementSecondary
+        self.elementStaticWhite = elementStaticWhite
+        self.elementSuccess = elementSuccess
+        self.elementSuccessDefault = elementSuccessDefault
+        self.elementSuccessPressed = elementSuccessPressed
+        self.specialExtra01 = specialExtra01
+        self.specialExtra02 = specialExtra02
+        self.specialExtra03 = specialExtra03
+        self.specialExtra04 = specialExtra04
+        self.specialExtra05 = specialExtra05
+        self.specialExtra06 = specialExtra06
+        self.specialExtra07 = specialExtra07
+        self.specialExtra08 = specialExtra08
+        self.specialExtra09 = specialExtra09
+        self.specialExtra10 = specialExtra10
+        self.specialExtra11 = specialExtra11
+        self.specialExtra12 = specialExtra12
+        self.specialExtra13 = specialExtra13
+        self.specialExtra14 = specialExtra14
+        self.specialExtra15 = specialExtra15
+        self.specialExtra16 = specialExtra16
+        self.specialExtra17 = specialExtra17
+        self.specialExtra18 = specialExtra18
+        self.specialExtra19 = specialExtra19
+        self.specialExtra20 = specialExtra20
+        self.specialExtra21 = specialExtra21
+        self.specialExtra22 = specialExtra22
+        self.specialExtra23 = specialExtra23
+        self.specialExtra24 = specialExtra24
+        self.specialExtra25 = specialExtra25
+        self.specialExtra26 = specialExtra26
+        self.specialExtra27 = specialExtra27
+        self.textAccent = textAccent
+        self.textAccentAdditional = textAccentAdditional
+        self.textAccentPressed = textAccentPressed
+        self.textAdditional = textAdditional
+        self.textAttention = textAttention
+        self.textAttentionDefault = textAttentionDefault
+        self.textAttentionPressed = textAttentionPressed
+        self.textContrast = textContrast
+        self.textError = textError
+        self.textErrorDefault = textErrorDefault
+        self.textErrorPressed = textErrorPressed
+        self.textMask = textMask
+        self.textPrimary = textPrimary
+        self.textSecondary = textSecondary
+        self.textStaticWhite = textStaticWhite
+        self.textSuccess = textSuccess
+        self.textSuccessDefault = textSuccessDefault
         self.textSuccessPressed = textSuccessPressed
     }
 
@@ -294,9 +304,9 @@ public class ColorPalette: Codable {
 
 // MARK: - All Fonts
 
-extension ColorPalette {
+public extension ColorPalette {
 
-    public var all: [AColor] {
+    var all: [AColor] {
         return [
             backgroundAccent,
             backgroundAccentDark,
@@ -388,12 +398,12 @@ extension ColorPalette {
 
 }
 
- 
+
 // MARK: - Dark
 
-extension ColorPalette {
+public extension ColorPalette {
 
-    public static var dark: ColorPalette {
+    static var dark: ColorPalette {
         return ColorPalette(
             style: .dark,
             disabledAlpha: 0.6,
@@ -487,12 +497,12 @@ extension ColorPalette {
     }
 
 }
- 
+
 // MARK: - Light
 
-extension ColorPalette {
+public extension ColorPalette {
 
-    public static var light: ColorPalette {
+    static var light: ColorPalette {
         return ColorPalette(
             style: .light,
             disabledAlpha: 0.6,
@@ -586,12 +596,12 @@ extension ColorPalette {
     }
 
 }
- 
+
 // MARK: - Smedark
 
-extension ColorPalette {
+public extension ColorPalette {
 
-    public static var sMEDark: ColorPalette {
+    static var sMEDark: ColorPalette {
         return ColorPalette(
             style: .dark,
             disabledAlpha: 0.6,
@@ -685,12 +695,12 @@ extension ColorPalette {
     }
 
 }
- 
+
 // MARK: - Smelight
 
-extension ColorPalette {
+public extension ColorPalette {
 
-    public static var sMELight: ColorPalette {
+    static var sMELight: ColorPalette {
         return ColorPalette(
             style: .light,
             disabledAlpha: 0.6,
@@ -783,110 +793,6 @@ extension ColorPalette {
         )
     }
 
-}
-
-// MARK: - Color
-
-public struct AColor: Codable {
-    public let red: CGFloat
-    public let green: CGFloat
-    public let blue: CGFloat
-    public let alpha: CGFloat
-    
-    public var uiColor: UIColor {
-        return UIColor(red: red, green: green, blue: blue, alpha: alpha)
-    }
-
-    public var cgColor: CGColor {
-        return uiColor.cgColor
-    }
-    
-    public static var clear: AColor {
-        return AColor(red: 0, green: 0, blue: 0, alpha: 0.0)
-    }
-    
-    public init(red: CGFloat, green: CGFloat, blue: CGFloat, alpha: CGFloat) {
-        self.red = red
-        self.green = green
-        self.blue = blue
-        self.alpha = alpha
-    }
-    
-    public init(red: Int, green: Int, blue: Int) {
-        self.init(red: CGFloat(red) / 255.0,
-                  green: CGFloat(green) / 255.0,
-                  blue: CGFloat(blue) / 255.0,
-                  alpha: 1.0)
-    }
-    
-    public init(hexString: String) {
-        let hex = hexString.trimmingCharacters(in: CharacterSet.alphanumerics.inverted)
-        var int = UInt64()
-        Scanner(string: hex).scanHexInt64(&int)
-        let a, r, g, b: UInt64
-        switch hex.count {
-        case 3: // RGB (12-bit)
-            (a, r, g, b) = (255, (int >> 8) * 17, (int >> 4 & 0xF) * 17, (int & 0xF) * 17)
-        case 6: // RGB (24-bit)
-            (a, r, g, b) = (255, int >> 16, int >> 8 & 0xFF, int & 0xFF)
-        case 8: // ARGB (32-bit)
-            (a, r, g, b) = (int >> 24, int >> 16 & 0xFF, int >> 8 & 0xFF, int & 0xFF)
-        default:
-            (a, r, g, b) = (255, 0, 0, 0)
-        }
-        
-        self.init(red: CGFloat(r) / 255, green: CGFloat(g) / 255, blue: CGFloat(b) / 255, alpha: CGFloat(a) / 255)
-    }
-    
-    public init(color: UIColor) {
-        let colorTo = CGColorSpace(name: CGColorSpace.sRGB)!
-        let cgColorInRGB = color.cgColor.converted(
-            to: colorTo,
-            intent: .defaultIntent,
-            options: nil)!
-        
-        let colorRef = cgColorInRGB.components
-        let red = colorRef?[0] ?? 0
-        let green = colorRef?[1] ?? 0
-        let blue = ((colorRef?.count ?? 0) > 2 ? colorRef?[2] : green) ?? 0
-        let alpha = color.cgColor.alpha
-        self.init(red: red, green: green, blue: blue, alpha: alpha)
-    }
-    
-    public func hexString(withAplha: Bool = true) -> String {
-        var colorString = withAplha ? String(format: "#%02lX", lroundf(Float(alpha * 255))) : "#"
-
-        colorString += String(
-            format: "%02lX%02lX%02lX",
-            lroundf(Float(red * 255)),
-            lroundf(Float(green * 255)),
-            lroundf(Float(blue * 255))
-        )
-    
-        return colorString
-    }
-    
-    public func withAlpha(_ alpha: CGFloat) -> AColor {
-        return AColor(red: red, green: green, blue: blue, alpha: alpha)
-    }
-    
-}
-
-@available(iOS 13.0.0, *)
-extension AColor {
-    
-    public var swiftUIColor: SwiftUI.Color {
-        return SwiftUI.Color(.sRGB, red: Double(red), green: Double(green), blue: Double(blue), opacity: Double(alpha))
-    }
-    
-}
-
-extension UIColor {
-    
-    public var color: AColor {
-        return AColor(color: self)
-    }
-    
 }
 
 // swiftlint:enable all
