+++ Source/AdmiralTheme/Classes/CoreFoundation/AppTheme/Fonts/FontPalette.swift
@@ -5,15 +5,25 @@
 //  Generated by Fastlane Figma Plugin.
 //
 
+import AdmiralUIResources
 import UIKit
-import SwiftUI
 
 // swiftlint:disable all
+/**
+ ColorPalette - The container that include a list of fonts.
 
-// MARK: - FontPalette
+ You can create your own FontPalette or  expand the current color palette by using the mechanism of inheritance
+ ## Example to create FontPalette
+ # Code
+ let colorPalette = FontPalette(...)
+ class CustomFontPalette: FontPalette {}
+ ```
+ */
+/// The container that include a list of font.
+open class FontPalette: Codable {
 
-public class FontPalette: Codable {
-    
+    // MARK: - Public Properties
+
     public var body1: AFont
     public var body2: AFont
     public var caption1: AFont
@@ -30,6 +40,8 @@ public class FontPalette: Codable {
     public var subtitle3: AFont
     public var title1: AFont
     public var title2: AFont
+
+    // MARK: - Initializer
 
     public init( 
         body1: AFont, 
@@ -71,9 +83,9 @@ public class FontPalette: Codable {
 
 // MARK: - All Fonts
 
-extension FontPalette {
+public extension FontPalette {
 
-    public var all: [AFont] {
+    var all: [AFont] {
         return [
             body1,
             body2,
@@ -96,12 +108,11 @@ extension FontPalette {
 
 }
 
- 
 // MARK: - Default
 
-extension FontPalette {
+public extension FontPalette {
 
-    public static var  `default`: FontPalette {
+    static var  `default`: FontPalette {
         return FontPalette(
             body1: AFont(name: "SFProDisplay-Medium", family: "SF Pro Display", size: 16.0, weight: 500, letterSpacing: 0.0, textStyle: .body),
             body2: AFont(name: "SFProDisplay-Regular", family: "SF Pro Display", size: 16.0, weight: 400, letterSpacing: 0.0, textStyle: .body),
@@ -124,52 +135,14 @@ extension FontPalette {
 
 }
 
+// MARK: - Extensions
 
-// MARK: - Font
+public extension FontPalette {
 
-public class AFont: Codable {
-    
-    public let name: String
-    public let family: String
-    public let size: CGFloat
-    public let weight: CGFloat
-    public let letterSpacing: CGFloat
-    public let textStyle: UIFont.TextStyle?
-    
-    public var uiFont: UIFont {
-        if let font = UIFont(name: name, size: size) {
-            return font
-        } else {
-            assert(true, "Font with name \"\(name)\" not found!")
-            return UIFont.systemFont(ofSize: size)
-        }
+    static func registerDefaultFonts() {
+        FontFamily.registerAllCustomFonts()
     }
-    
-    public init(
-        name: String,
-        family: String,
-        size: CGFloat,
-        weight: CGFloat,
-        letterSpacing: CGFloat,
-        textStyle: UIFont.TextStyle?
-    ) {
-        self.name = name
-        self.family = family
-        self.size = size
-        self.weight = weight
-        self.letterSpacing = letterSpacing
-        self.textStyle = textStyle
-    }
-    
-}
 
-@available(iOS 13.0.0, *)
-extension AFont {
-    
-    public var swiftUIFont: SwiftUI.Font {
-        return SwiftUI.Font(self.uiFont)
-    }
-    
 }
 
 extension UIFont.TextStyle: Codable {}
