--- Source/AdmiralUIResources/Classes/CoreFoundation/Fonts/Fonts+Register.swift
@@ -1,16 +1,0 @@
-//
-//  Fonts+Register.swift
-//  AdmiralUIResources
-//
-//  Created on 03.04.2021.
-//
-
-import Foundation
-
-extension FontPalette {
-    
-    public static func registerDefaultFonts() {
-        FontFamily.registerAllCustomFonts()
-    }
-    
-}
