--- Source/AdmiralUIResources/Classes/Extensions/UIKit/UIFont/Fonts+Registration.swift
@@ -1,39 +1,0 @@
-//
-//  Fonts+Registration.swift
-//  AdmiralUI
-//
-//  Created on 09.11.2020.
-//
-
-import UIKit
-
-public extension UIFont {
-    
-    class func registerFontWithFilenameString(filenameString: String, bundle: Bundle) {
-        guard let pathForResourceString = bundle.path(forResource: filenameString, ofType: nil) else {
-            print("UIFont+:  Failed to register font - path for resource not found.")
-            return
-        }
-        
-        guard let fontData = NSData(contentsOfFile: pathForResourceString) else {
-            print("UIFont+:  Failed to register font - font data could not be loaded.")
-            return
-        }
-        
-        guard let dataProvider = CGDataProvider(data: fontData) else {
-            print("UIFont+:  Failed to register font - data provider could not be loaded.")
-            return
-        }
-        
-        guard let fontRef = CGFont(dataProvider) else {
-            print("UIFont+:  Failed to register font - font could not be loaded.")
-            return
-        }
-        
-        var errorRef: Unmanaged<CFError>?
-        if CTFontManagerRegisterGraphicsFont(fontRef, &errorRef) == false {
-            print("UIFont+:  Failed to register font - register graphics font failed - this font may have already been registered in the main bundle.")
-        }
-    }
-    
-}
